refactor: replace strings with constants, update filenames with -tmp suffix, align error message with language used
Replaced strings with constants to improve maintainability. Added the suffix `-tmp` to filenames to handle case sensitivity issues in the repository. Updated an error message to align with the language used.
