+++ src/components/button/Button-tmp.jsx
@@ -1,4 +1,4 @@
-import './button.css';
+import './Button-tmp.css';
 
 export default function Button ({ onClick, isDisabled , children}) {
   return (
