+++ src/services/questions/getQuestions.js
@@ -2,7 +2,7 @@ export const getQuestions = async () => {
   try {
     const response = await fetch('http://localhost:3000/questions');
     if (!response.ok) {
-      throw new Error('Network response was not ok');
+      throw new Error('La respuesta del servidor no fue favorable');
     }
     const data = await response.json();
     return data;
