+++ src/utils/textConstant.js
@@ -1,0 +1,6 @@
+export const SEMANA = 'Semana';
+export const DIA = 'Dia';
+export const MAXIMO = 'Maximo';
+export const CARACTERES = 'Caracteres';
+export const TIPO_DE_PLAN = 'Tipo de plan: ';
+export const DIAS_POR_SEMANA = 'Días por semana: ';
